remove conversion to double tensor

+++ model/model_p/tcn_util.py
@@ -172,5 +172,6 @@ class TCN_Classifier(nn.Module):
 		"""
 		out_embedding = self.tcn(x)
 		# out = self.linear(out_embedding[:, :, -1])
-		out = self.linear(out_embedding).double()
+		# out = self.linear(out_embedding).double()
+		out = self.linear(out_embedding)
 		return self.output(out)
